Add sorting of guests alphebatically.
Added the `System.Linq` namespace to `Program.cs` to enable LINQ methods. Modified the guest report generation to sort guests alphabetically by last name and then by first name, enhancing the readability and organisation of the output. This change fulfills the request within the requirements for the lists to be ordered alphabetically.